Pre code fix

+++ app/src/main/java/com/platform45/weather45/models/SeriesDateData.kt
@@ -3,8 +3,10 @@ package com.platform45.weather45.models
 import com.google.gson.annotations.SerializedName
 
 data class SeriesDateData(
-    @SerializedName("close") var close: Double?,
-    @SerializedName("high") var high: Double?,
-    @SerializedName("low") var low: Double?,
-    @SerializedName("open") var open: Double?
+    @SerializedName("close") var close: Float?,
+    @SerializedName("high") var high: Float?,
+    @SerializedName("low") var low: Float?,
+    @SerializedName("open") var open: Float?
 )
+
+
